Ajustado para não dar erro nas fk que não encontra e a busca por todas as sprints

+++ src/domain/dto/auxiliar/progressTask/progressTask.go
@@ -1,6 +1,6 @@
 package progressTaskDto
 
 type ProgressTask struct {
-	TaskId     int64
-	ProgressId int64
+	TaskId     uint
+	ProgressId uint
 }
