+++ src/domain/dto/auxiliar/statusTask/statusTask.go
@@ -1,6 +1,6 @@
 package statusTaskDto
 
 type StatusTask struct {
-	TaskId   int64
-	StatusId int64
+	TaskId   uint
+	StatusId uint
 }
